feat: int타입을 long으로 변경

+++ android/Signal/app/src/main/java/com/ongo/signal/data/repository/main/comment/CommentRepository.kt
@@ -8,7 +8,7 @@ import retrofit2.Response
 
 interface CommentRepository {
 
-    suspend fun readComments(boardId: Int): Response<CommentDTO>
+    suspend fun readComments(boardId: Long): Response<CommentDTO>
     suspend fun writeComment(boardId: Long, commentDto: CommentDTOItem): Response<CommentDTOItem>
     suspend fun updateComment(
         boardId: Long,
